CNCMicroscope: Eliminate VC++ warning for loss of precision
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@15689 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/CNCMicroscope/ArduinoNeoPixelShutter/ArduinoNeoPixelShutter.h
@@ -59,7 +59,10 @@ public:
    bool IsPortAvailable() {return portAvailable_;}
 
    int PurgeComPortH() {return PurgeComPort(port_.c_str());}
-   int WriteToComPortH(const unsigned char* command, unsigned len) {return WriteToComPort(port_.c_str(), command, len);}
+   int WriteToComPortH(const unsigned char* command, size_t len)
+   {
+      return WriteToComPort(port_.c_str(), command, static_cast<unsigned>(len));
+   }
    int ReadFromComPortH(unsigned char* answer, unsigned maxLen, unsigned long& bytesRead)
    {
       return ReadFromComPort(port_.c_str(), answer, maxLen, bytesRead);
